pyproject.toml
[project]
name = "laminar-examples"
version = "0.1.0"
description = "Proven integration examples for Laminar AI observability at OrbitronAI"
requires-python = ">=3.10"
dependencies = [
    "lmnr>=0.4",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "openai>=1.0",
    "httpx[http2]>=0.27",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
